rename buyasset to downloadasset
