--- client/src/components/+buy-asset/buy-asset.module.scss
@@ -1,5 +1,0 @@
-@import '~@nevermined-io/styles/lib/cjs/styles/index.scss';
-
-.buy-asset {
-  cursor: pointer;
-}
