--- client/src/components/+buy-asset/popup/buy-asset-popup.module.scss
@@ -1,32 +1,0 @@
-@import '~@nevermined-io/styles/lib/cjs/styles/index.scss';
-
-.buy-asset-popup {
-  color: gc(gD);
-
-  &__confirm {
-    width: 512px;
-    height: 320px;
-    box-shadow: 0px 10px 30px rgba(0, 0, 0, 0.1);
-    padding:40px 15px 40px 15px;
-  }
-
-  &__text {
-    color: gc(gD) !important;
-
-    &--content {
-      max-width: 340px;
-      margin: auto;
-    }
-  }
-  &__button {
-    flex: 1;
-  }
-  &__icon {
-    &--error {
-      color: gc(error)
-    }
-    &--success {
-      color: gc(success)
-    }
-  }
-}
